+++ lib/nugrant/vagrant/v2/plugin.rb
@@ -1,4 +1,4 @@
-
+require 'nugrant/vagrant/v2/action'
 module Nugrant
   module Vagrant
     module V2
@@ -7,6 +7,16 @@ module Nugrant
         description <<-DESC
           Plugin to define and use user specific parameters from various location inside your Vagrantfile.
         DESC
+
+        class << self
+          def provision(hook)
+            hook.before(::Vagrant::Action::Builtin::Provision, Nugrant::Vagrant::V2::Action.autoExport)
+          end
+        end
+
+        action_hook(:nugrant_provision, :machine_action_up, &method(:provision))
+        action_hook(:nugrant_provision, :machine_action_reload, &method(:provision))
+        action_hook(:nugrant_provision, :machine_action_provision, &method(:provision))
 
         command "user" do
           require_relative "command/root"