+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # 2.0.3 (In progress)
+* Added possibility to automatically export variables on provision to env file.
+  This can be used in Vagrant by doing `config.user.auto_export = :format`
+  where valid format are:
 
+   * false => Default value, disable auto export
+   * :autoenv => Export in [autoenv] format
+   * :script => Export to a script file
+     * The default generated script path is "./nugrant2env.sh". 
+       Change it by doing in Vagrant
+       `config.user.auto_export_script_path = "./script/exemple.sh`
+   * `[:autoenv, :script]` export both format
+         
+   
 # 2.0.2 (July 13th, 2014)
 
  * Fixed a bug when using some vagrant commands. The code to retrieve